requirements.txt
Flask==3.0.3
Jinja2==3.1.4
requests==2.32.3
websockets==12.0
solana==0.27.1  # solana-py
python-dotenv==1.0.1
SQLAlchemy==2.0.30
python-telegram-bot==21.0.1
discord.py==2.3.2
tweepy==4.14.0
